+++ plugins/core/actions/core.py
@@ -218,3 +218,5 @@ def tell_month(**kwargs):
     ]
 
     app.say(random.choice(phrases))
+
+